ref(jsdom): Move attribute query helpers to utility
Keep index focused on routing the querySelectorAll fast path and fallback. The traversal and namespace matching now live alongside the other tree utilities.
